style(PrizeCategory): long 타입으로 변환

+++ src/main/java/lotto/model/PrizeCategory.java
@@ -1,37 +1,37 @@
 package lotto.model;
 
 public enum PrizeCategory {
-    THREE_MATCH("3개 일치", 3, 5000) {
-        int calculatePrize(int ticketQuantity) {
+    THREE_MATCH("3개 일치", 3, 5000L) {
+        long calculatePrize(int ticketQuantity) {
             return ticketQuantity * BASIC_PRIZE;
         }
     },
-    FOUR_MATCH("4개 일치", 4, 50000) {
-        int calculatePrize(int ticketQuantity) {
+    FOUR_MATCH("4개 일치", 4, 50000L) {
+        long calculatePrize(int ticketQuantity) {
             return ticketQuantity * BASIC_PRIZE;
         }
     },
-    FIVE_MATCH_NO_BONUS("5개 일치", 5, 1500000) {
-        int calculatePrize(int ticketQuantity) {
+    FIVE_MATCH_NO_BONUS("5개 일치", 5, 1500000L) {
+        long calculatePrize(int ticketQuantity) {
             return ticketQuantity * BASIC_PRIZE;
         }
     },
-    FIVE_MATCH_WITH_BONUS("5개 일치, 보너스 볼 일치", 5, 30000000) {
-        int calculatePrize(int ticketQuantity) {
+    FIVE_MATCH_WITH_BONUS("5개 일치, 보너스 볼 일치", 5, 30000000L) {
+        long calculatePrize(int ticketQuantity) {
             return ticketQuantity * BASIC_PRIZE;
         }
     },
-    SIX_MATCH("6개 일치", 6, 2000000000) {
-        int calculatePrize(int ticketQuantity) {
+    SIX_MATCH("6개 일치", 6, 2000000000L) {
+        long calculatePrize(int ticketQuantity) {
             return ticketQuantity * BASIC_PRIZE;
         }
     };
 
     protected final String NAME;
     protected final int MATCH_COUNT;
-    protected final int BASIC_PRIZE;
+    protected final long BASIC_PRIZE;
 
-    PrizeCategory(String name, int matchCount, int basicPrize) {
+    PrizeCategory(String name, int matchCount, long basicPrize) {
         this.NAME = name;
         this.MATCH_COUNT = matchCount;
         this.BASIC_PRIZE = basicPrize;
@@ -45,9 +45,9 @@ public enum PrizeCategory {
         return MATCH_COUNT;
     }
 
-    public int getBasicPrize() {
+    public long getBasicPrize() {
         return BASIC_PRIZE;
     }
 
-    abstract int calculatePrize(int ticketQuantity);
+    abstract long calculatePrize(int ticketQuantity);
 }
